pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "product-guide"
version = "0.1.0"
description = "美食/超市商品导购助手（CLI + Chroma RAG + 豆包 Ark）"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "chromadb>=0.5",
    "volcengine-python-sdk[ark]",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = ["pytest"]

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-dir]
"" = "src"
